nice bro

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: rmaanane <rmaanane@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/22 00:54:46 by rmaanane          #+#    #+#             */
-/*   Updated: 2024/11/05 05:59:01 by rmaanane         ###   ########.fr       */
+/*   Updated: 2024/11/06 03:33:56 by rmaanane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] && s2[i] && i < n)
